Renaming flexvolume to rookflex
Fixes https://github.com/rook/rook/issues/1052
