+++ cmd/rookflex/cmd/isattached.go
